+++ source/arrays.c
@@ -18,8 +18,8 @@ dynArray debts;
 dynArray investments;
 
 // Misc Portfolio Data
-float income;
-float assets;
+double income;
+double assets;
 
 void cleanupArray(dynArray *array) { // Reallocates a given array to only take up as much space as is in use
     array->size = array->used;
